+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Add [raw](https://docs.imgproxy.net/latest/generating_the_url?id=raw) processing option.
 - (pro) Add encrypted source URL support.
 
+### Changed
+- Fix some invalid signature cases that happen because of URL normalization.
+
 ## [3.7.2] - 2022-08-22
 ### Changed
 - (docker) Faster images quantization.